refactor pack filtering
